infra/mcp-server-hub/src/vertex_ai_integration.py
#!/usr/bin/env python3

import os
import json
import logging
from typing import Dict, List, Optional, Any
from google.cloud import aiplatform
from vertexai.generative_models import GenerativeModel, SafetySetting, HarmCategory

class VertexAIAgentCoordinator:
    """
    Integrates Vertex AI capabilities with MCP Server Hub for intelligent agent coordination
    """
    
    def __init__(self, project_id: str, location: str = "us-central1", logger=None):
        self.project_id = project_id
        self.location = location
        self.logger = logger or logging.getLogger(__name__)
        
        # Initialize Vertex AI
        aiplatform.init(project=project_id, location=location)
        
        # Initialize Generative Model
        self.model = GenerativeModel("gemini-pro")
        
        # Safety settings for agent interactions
        self.safety_settings = [
            SafetySetting(
                category=HarmCategory.HARM_CATEGORY_HATE_SPEECH,
                threshold=SafetySetting.HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE
            ),
            SafetySetting(
                category=HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT,
                threshold=SafetySetting.HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE
            )
        ]
    
    async def intelligent_tool_selection(self, user_request: str, available_tools: List[Dict]) -> Dict[str, Any]:
        """
        Use Vertex AI to intelligently select and configure tools for a user request
        """
        try:
            # Prepare tool descriptions for the model
            tools_description = "\n".join([
                f"- {tool['name']}: {tool.get('description', 'No description')}"
                for tool in available_tools
            ])
            
            prompt = f"""
            Given this user request: "{user_request}"
            
            Available tools:
            {tools_description}
            
            Select the most appropriate tool(s) and generate the parameters needed.
            
            Respond in JSON format:
            {{
                "selected_tool": "tool_name",
                "parameters": {{"key": "value"}},
                "reasoning": "explanation of choice",
                "confidence": 0.95
            }}
            """
            
            response = self.model.generate_content(
                prompt,
                safety_settings=self.safety_settings
            )
            
            # Parse the JSON response
            result = json.loads(response.text)
            
            self.logger.info(f"AI selected tool: {result.get('selected_tool')} with confidence: {result.get('confidence')}")
            
            return result
            
        except Exception as e:
            self.logger.error(f"Error in intelligent tool selection: {e}")
            return {
                "selected_tool": None,
                "parameters": {},
                "reasoning": f"Error: {str(e)}",
                "confidence": 0.0
            }
    
    async def coordinate_multi_agent_workflow(self, task_description: str, available_agents: List[Dict]) -> Dict[str, Any]:
        """
        Use Vertex AI to orchestrate multi-agent workflows
        """
        try:
            agents_description = "\n".join([
                f"- {agent['name']}: {agent.get('description', 'No description')}"
                for agent in available_agents
            ])
            
            prompt = f"""
            Task: {task_description}
            
            Available agents:
            {agents_description}
            
            Create an optimal workflow that coordinates these agents to complete the task.
            
            Respond in JSON format:
            {{
                "workflow_steps": [
                    {{
                        "step": 1,
                        "agent": "agent_name",
                        "action": "specific action",
                        "dependencies": ["previous_step_ids"],
                        "estimated_duration": "5 minutes"
                    }}
                ],
                "coordination_strategy": "parallel|sequential|hybrid",
                "risk_assessment": "low|medium|high",
                "expected_outcome": "description"
            }}
            """
            
            response = self.model.generate_content(
                prompt,
                safety_settings=self.safety_settings
            )
            
            workflow = json.loads(response.text)
            
            self.logger.info(f"Generated workflow with {len(workflow.get('workflow_steps', []))} steps")
            
            return workflow
            
        except Exception as e:
            self.logger.error(f"Error in workflow coordination: {e}")
            return {
                "workflow_steps": [],
                "coordination_strategy": "error",
                "risk_assessment": "high",
                "expected_outcome": f"Failed to generate workflow: {str(e)}"
            }
    
    async def enhance_a2a_message_conversion(self, a2a_message: str, available_mcp_tools: List[Dict]) -> Dict[str, Any]:
        """
        Use Vertex AI to intelligently convert A2A messages to MCP tool calls
        """
        try:
            tools_info = "\n".join([
                f"- {tool['name']}: {tool.get('description', '')} (parameters: {tool.get('parameters', {})})"
                for tool in available_mcp_tools
            ])
            
            prompt = f"""
            Convert this A2A agent message to an appropriate MCP tool call:
            Message: "{a2a_message}"
            
            Available MCP tools:
            {tools_info}
            
            Respond in JSON format:
            {{
                "tool_name": "selected_tool",
                "parameters": {{"param1": "value1"}},
                "conversion_confidence": 0.95,
                "alternative_tools": ["tool2", "tool3"],
                "interpretation": "how you understood the message"
            }}
            """
            
            response = self.model.generate_content(
                prompt,
                safety_settings=self.safety_settings
            )
            
            conversion = json.loads(response.text)
            
            self.logger.info(f"Converted A2A message to MCP tool: {conversion.get('tool_name')}")
            
            return conversion
            
        except Exception as e:
            self.logger.error(f"Error in A2A message conversion: {e}")
            return {
                "tool_name": None,
                "parameters": {},
                "conversion_confidence": 0.0,
                "alternative_tools": [],
                "interpretation": f"Conversion failed: {str(e)}"
            }
    
    async def generate_agent_response(self, context: str, agent_capabilities: List[str]) -> str:
        """
        Generate intelligent responses for agents based on context and capabilities
        """
        try:
            capabilities_text = ", ".join(agent_capabilities)
            
            prompt = f"""
            You are an AI agent with these capabilities: {capabilities_text}
            
            Context: {context}
            
            Generate an appropriate response that demonstrates your capabilities and provides value.
            Keep the response concise and actionable.
            """
            
            response = self.model.generate_content(
                prompt,
                safety_settings=self.safety_settings
            )
            
            return response.text
            
        except Exception as e:
            self.logger.error(f"Error generating agent response: {e}")
            return f"I encountered an error while processing your request: {str(e)}"
    
    def validate_agent_interaction_safety(self, interaction_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Assess safety and compliance of agent interactions
        """
        try:
            # Use Vertex AI safety features to assess the interaction
            content = str(interaction_data)
            
            # This is a simplified example - in practice, you'd implement more sophisticated safety checks
            safety_assessment = {
                "is_safe": True,
                "risk_level": "low",
                "concerns": [],
                "recommendations": []
            }
            
            return safety_assessment
            
        except Exception as e:
            self.logger.error(f"Error in safety validation: {e}")
            return {
                "is_safe": False,
                "risk_level": "unknown",
                "concerns": [f"Safety validation failed: {str(e)}"],
                "recommendations": ["Manual review required"]
            }