infra/mcp-server-hub/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
pydantic>=2.5.0
mcp>=1.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-cloud-aiplatform>=1.38.0
vertexai>=0.1.0